backend/requirements.txt
# FastAPI and web framework dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Pydantic for data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI/ML dependencies
torch>=2.6.0
torchvision>=0.21.0
numpy>=1.24.3
Pillow>=10.1.0
opencv-python>=4.8.1.78

# HTTP client for external APIs
httpx>=0.25.2
aiohttp>=3.9.1

# Database and caching
asyncpg>=0.29.0
redis>=5.0.1
sqlalchemy>=2.0.23
alembic>=1.13.1

# Environment and configuration
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.2.0
psutil>=5.9.6

# Production dependencies only
# Development dependencies are in requirements-dev.txt